requirements.txt
torch
torchvision
numpy
pillow
tqdm
matplotlib
albumentations
scikit-learn
shap
lime
tensorboard
